Implementando loanedBookErrorOnCreateLoanTest().
