+++ src/main/java/com/cursodsousa/libraryapi/api/ApplicationControllerAdvice.java
@@ -1,4 +1,34 @@
 package com.cursodsousa.libraryapi.api;
 
+import com.cursodsousa.libraryapi.api.exception.ApiErros;
+import com.cursodsousa.libraryapi.exception.BusinessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.server.ResponseStatusException;
+
+@RestControllerAdvice
 public class ApplicationControllerAdvice {
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ApiErros handleValidationExceptions(MethodArgumentNotValidException ex){ //Tratando Exception - Mapeia uma exception para um retorno
+        BindingResult bindingResult = ex.getBindingResult();
+        return new ApiErros(bindingResult);
+    }
+
+    @ExceptionHandler(BusinessException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ApiErros handleBusinessException(BusinessException ex){
+        return new ApiErros(ex);
+    }
+
+    @ExceptionHandler(ResponseStatusException.class)
+    public ResponseEntity handleResponseStatusException(ResponseStatusException ex ){
+        return new ResponseEntity(new ApiErros(ex), ex.getStatus());
+    }
 }
