+++ src/main/java/com/cursodsousa/libraryapi/api/exception/ApiErros.java
@@ -2,6 +2,7 @@ package com.cursodsousa.libraryapi.api.exception;
 
 import com.cursodsousa.libraryapi.exception.BusinessException;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -17,6 +18,10 @@ public class ApiErros {
 
     public ApiErros(BusinessException ex){
         this.errors = Arrays.asList(ex.getMessage());
+    }
+
+    public ApiErros(ResponseStatusException ex){
+        this.errors = Arrays.asList(ex.getReason());
     }
 
     public List<String> getErrors(){
